fix-2: move fetch getProducts to ProductPage
