+++ src/router.js
@@ -1,6 +1,5 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { API_URL } from "./constants/constants";
 import "./router.css";
 import Login from "./pages/Login/Login";
 import ProductsPage from "./pages/ProductsPage/ProductsPage";
@@ -11,31 +10,6 @@ import PreviewPage from "./pages/PreviewPage/PreviewPage";
 
 
 const RouterApp = () => {
-  const [products, setProducts] = useState([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [isError, setIsError] = useState(false);
-
-
-  useEffect(() => {
-    getProducts();
-  }, []);
-
-  const getProducts = async () => {
-    try {
-      const response = await fetch(API_URL + "products");
-
-      if (!response.ok) {
-        throw new Error("Something Error");
-      }
-
-      const productsData = await response.json();
-      setProducts(productsData);
-      setIsLoading(false);
-    } catch (error) {
-      setIsError(true);
-      setIsLoading(false);
-    }
-  };
 
   return (
     <div className="App">
@@ -45,7 +19,7 @@ const RouterApp = () => {
           <Route path="/login" element={<Login />} />
 
         <Route element={<PrivateRoute/>} >
-          <Route path="/product-page" element={<ProductsPage products={products} isLoading={isLoading} isError={isError}/>} />
+          <Route path="/product-page" element={<ProductsPage />} />
           <Route path="/preview-page" element={<PreviewPage />} />
           <Route path="/product-preview/:id" element={<ProductPreview />} />
         </Route>
